graph randomization is now its own module
